backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
python-dotenv==1.2.1
psycopg2-binary==2.9.11

django-allauth[headless]==65.13.1
dj-rest-auth==7.0.1

google-api-python-client==2.187.0
google-auth==2.34.0
google-auth-oauthlib==1.2.3

pytest-django==4.11.1